Imported Upstream version 3.14.11

+++ src/libCom/osi/os/RTEMS/osdInterrupt.c
@@ -6,7 +6,7 @@
 \*************************************************************************/
 /*
  * RTEMS osdInterrupt.c
- *	osdInterrupt.c,v 1.6.2.1 2007/12/17 18:43:59 norume Exp
+ *	osdInterrupt.c,v 1.6.2.2 2009/04/24 17:01:39 anj Exp
  *      Author: W. Eric Norum
  *              eric@cls.usask.ca
  *              (306) 966-6055
@@ -65,7 +65,7 @@ void
 InterruptContextMessageDaemon (void *unused)
 {
     const char *message;
-    uint32_t size;
+    size_t size;
     rtems_status_code sc;
 
     sc = rtems_message_queue_create (rtems_build_name ('I', 'C', 'M', 'Q'),
